pyproject.toml
[project]
name = "fansme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boltons>=24.1.0",
    "fastapi-sso>=0.17.0",
    "fastapi[standard]>=0.115.5",
    "ipython>=8.31.0",
    "mongoengine>=0.29.1",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.10.1",
    "redis>=5.2.0",
    "types-redis>=4.6.0.20241004",
    "uvicorn>=0.32.0",
]
